Set integration bridge in localrc
XEN_INTEGRATION_BRIDGE should point to the dom0 bridge that serves the
purpose of the integration bridge. The value will be taken from the
DEVSTACK_GATE_XS_INT_BR parameter. In normal cases this value is
injected through kernel parameters, however we don't have that
information in the cloud based OSCI.